Restructure to use dynamic import
