--- pages/_app.old
@@ -1,7 +1,0 @@
-// import '../styles/globals.css'
-// import type { AppProps } from 'next/app'
-// import { AblyProvider } from "@ably-labs/react-hooks"
-
-// export default function App({ Component, pageProps }: AppProps) {
-//   return <Component {...pageProps} />
-// }
